refactor(server): extract data mover crate
