+++ Cargo.toml
@@ -32,6 +32,7 @@ members = [
     "crates/adapters/curvine-ufs-opendal",
     "crates/adapters/curvine-ufs-oss-hdfs",
     "crates/adapters/curvine-hdfs-jni",
+    "crates/server/curvine-data-mover",
     "curvine-master",
     "curvine-worker",
     "curvine-common",
@@ -74,6 +75,7 @@ default-members = [
     "crates/infra/curvine-io",
     "crates/metadata/curvine-raft",
     "crates/adapters/curvine-storage-local",
+    "crates/server/curvine-data-mover",
     "curvine-master",
     "curvine-worker",
     "curvine-common",
@@ -112,6 +114,7 @@ curvine-storage-spdk = { path = "crates/adapters/curvine-storage-spdk" }
 curvine-ufs-opendal = { path = "crates/adapters/curvine-ufs-opendal" }
 curvine-ufs-oss-hdfs = { path = "crates/adapters/curvine-ufs-oss-hdfs" }
 curvine-hdfs-jni = { path = "crates/adapters/curvine-hdfs-jni" }
+curvine-data-mover = { path = "crates/server/curvine-data-mover" }
 curvine-fault = { path = "crates/infra/curvine-fault" }
 curvine-error = { path = "crates/common/curvine-error" }
 curvine-proto = { path = "crates/common/curvine-proto" }
