+++ crates/server/curvine-data-mover/src/common/ufs_client.rs
@@ -14,12 +14,12 @@
 
 #![allow(unused)]
 
-use curvine_client::unified::{UfsFileSystem, UnifiedReader, UnifiedWriter};
 use curvine_common::error::FsError;
 use curvine_common::fs::{CurvineURI, FileSystem, Path};
 use curvine_common::state::FileStatus;
 use curvine_common::FsResult;
 use curvine_ufs_api::fs::ufs_context::UFSContext;
+use curvine_unified_fs::{UfsFileSystem, UnifiedReader, UnifiedWriter};
 use std::sync::Arc;
 
 #[derive(Clone)]
