+++ crates/server/curvine-data-mover/src/lib.rs
@@ -12,4 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use curvine_data_mover::common::UfsFactory;
+pub mod common;
+pub mod transfer;
+pub mod worker;
+
+mod job_worker_client;
+pub use self::job_worker_client::JobWorkerClient;
+
+mod rpc_context;
+pub use self::rpc_context::RpcContext;
