+++ crates/server/curvine-data-mover/src/worker/mod.rs
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use curvine_data_mover::common::UfsFactory;
+pub mod task;
