+++ curvine-master/src/master/job/mod.rs
@@ -18,8 +18,7 @@ pub use crate::master::job::job_manager::JobManager;
 mod job_handler;
 pub use job_handler::JobHandler;
 
-mod job_worker_client;
-pub use self::job_worker_client::JobWorkerClient;
+pub use curvine_data_mover::JobWorkerClient;
 
 mod job_store;
 pub use job_store::JobStore;
