+++ curvine-worker/src/worker/task/mod.rs
@@ -12,17 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Worker side data loading module
-//! Responsible for loading data from external storage systems (such as S3) to local storage
-
-mod load_task_runner;
-pub use self::load_task_runner::LoadTaskRunner;
-
-mod task_manager;
-pub use self::task_manager::TaskManager;
-
-mod task_store;
-pub use self::task_store::TaskStore;
-
-mod task_context;
-pub use self::task_context::TaskContext;
+pub use curvine_data_mover::worker::task::*;
